Implemented parsing of reference PCR values from a text file.
The text file is expected to be in the same format as the output of 'tpm2 pcrread', e.g.:
0 : 0x0000000000000000000000000000000000000000000000000000000000000000
...
23: 0x0000000000000000000000000000000000000000000000000000000000000000
Entries are identified by the number at the start of the line. Entries are allowed to be missing if they are not in the reference_pcr_selection. Entries are expected to be in order.

By default the PCRs are read from the file reference-pcrs.txt, alternative paths can be passed via the CLI.

Signed-off-by: Dominik Lorych <dominik.lorych@sit.fraunhofer.de>