Added DTLS support based on RPK (raw public keys). This is intended for scenarios where attester and verifier have pre-shared their public keys with each other. The protocol only accepts connections from peers using these pre-shared keys. Alternatively the verification of the peers' public key also can be disabled, this is primarily intended for the verifier which may not know the public keys of all attesters and is able to check the identity with the attestation response. This commit includes default keys, but these are only intended for testing. They MUST be changed in actual production environments!
This commit also uses specific commits of libcoap and mbedtls in the Dockerfile as the current heads of their branches are not compatible with Charra. It also includes a patch file for a libcoap bug. This bug will be fixed in a future version of libcoap. https://github.com/obgm/libcoap/pull/671

Signed-off-by: Dominik Lorych <dominik.lorych@sit.fraunhofer.de>